Core dumps on exit overspilled the barrel.  Temporary fixes for issues to be resolved properly later.
- Added error handlers for RAT ui and engine components.
- All ignore error if should_exit is TRUE.
- Control error handler only kills processes now if should_exit FALSE.
- Mapped mbus.bye message to same fn as mbus.quit.

+++ main_control.c
@@ -330,9 +330,11 @@ static void mbus_err_handler(int seqnum, int reason)
 	/* we don't try to recover, just kill off the media engine and user  */
 	/* interface processes and exit.                                     */
 	printf("FATAL ERROR: couldn't send mbus message (%d:%d)\n", seqnum, reason);
-	kill_process(pid_ui);
-	kill_process(pid_engine);
-	abort();
+        if (should_exit == FALSE) {
+                kill_process(pid_ui);
+                kill_process(pid_engine);
+                abort();
+        }
 }
 
 static void terminate(struct mbus *m, char *addr)
